backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# Environment Variables
python-decouple==3.8

# Password Hashing
argon2-cffi==23.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# API Documentation
drf-spectacular==0.27.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Celery (Background Tasks)
celery==5.3.4
redis==5.0.1

# Rate Limiting
django-ratelimit==4.1.0

# Security - File Type Validation (optional, falls back to extension check)
# python-magic==0.4.27

# Security - Encrypted Fields (optional, enable if FIELD_ENCRYPTION_KEY is set)
# django-encrypted-model-fields==0.6.5
# cryptography>=41.0.0

# Development
django-extensions==3.2.3
ipython==8.20.0
